Resolve some 64-bit build warnings. (dm)
git-svn-id: svn://mielke.cc/main/brltty@4285 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Drivers/Braille/EuroBraille/eu_bluetooth.c
@@ -64,8 +64,8 @@ eubrl_bluetoothWrite (BrailleDisplay *brl, const void *buf, size_t length)
 	} 
       else 
 	{
-	  LogPrint(LOG_WARNING, "Trunccated bluetooth write: %d < %d", 
-		   count, length);
+	  LogPrint(LOG_WARNING, "Trunccated bluetooth write: %d < %u", 
+		   (int)count, (unsigned int)length);
 	}
     }
   return count;
